Add sh suffix and fix shellcheck errors.

+++ install.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 
-#-----------------------------------------------------
+#----------------------------------------------------------------------------------------
+# Well documented, terminal centric web developer's dot files.
+# Debian/ubuntu based distros installer.
 # @author Martin Toma
+#
 # @version 6.0
 # @updated Sat Nov  7 22:38:01 CET 2015
 # @created Fri Nov 15 13:13:22 CET 2013
-#-----------------------------------------------------
+#----------------------------------------------------------------------------------------
 
 # Dont continue on error
 set -e
@@ -16,15 +19,15 @@ REPLACE_FILES=false
 #-----------------------------------------------------
 # Functions and variables
 #-----------------------------------------------------
-current_path=`pwd`
+current_path=$(pwd)
 
-command_exists () {
-  type "$1" > /dev/null 2>$1
+command_exists() {
+  type "$1" > "/dev/null 2>$1"
 }
 
-install_oh_my_zsh () {
+install_oh_my_zsh() {
   curl -L https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh | sh
-  ln -sf $current_path/shell/martinus.zsh-theme ~/.oh-my-zsh/themes/martinus.zsh-theme
+  ln -sf "$current_path/shell/martinus.zsh-theme ~/.oh-my-zsh/themes/martinus.zsh-theme"
   echo "    Change your default shell to zsh"
   sudo chsh
 }
@@ -36,17 +39,27 @@ install_plug_nvim() {
 install_nvim_folder() {
   mkdir -p ~/.config/nvim/autoload
   install_plug_nvim
-  ln -sf $current_path/neovim/spell/dictionary.utf-8.add ~/.config/nvim/dictionary.utf-8.add
-  ln -sf $current_path/neovim/UltiSnips ~/.config/nvim/UltiSnips
-  ln -sf $current_path/neovim/init.vim ~/.config/nvim/init.vim
+  ln -sf "$current_path/neovim/spell/dictionary.utf-8.add ~/.config/nvim/dictionary.utf-8.add"
+  ln -sf "$current_path/neovim/UltiSnips ~/.config/nvim/UltiSnips"
+  ln -sf "$current_path/neovim/init.vim ~/.config/nvim/init.vim"
 }
 
 #-----------------------------------------------------
 # Basic requirements check
 #-----------------------------------------------------
 
-if !command_exists curl; then
+if ! command_exists pt-get; then
+  echo "This istaller is only comaptible with debian/ubuntu based Linux distributrions."
+  echo "Please install configuration files manually."
+  exit
+fi
+
+if ! command_exists curl; then
   sudo apt-get install -y curl
+fi
+
+if ! command_exists git; then
+  sudo apt-get install -y git
 fi
 
 #-----------------------------------------------------
@@ -56,11 +69,11 @@ echo -n "[ zshrc ]"
 
 if [ ! -f ~/.zshrc ]; then
   echo "    Creating zshrc!"
-  ln -sf $current_path/shell/zshrc ~/.zshrc
-else if REPLACE_FILES; then
+  ln -sf "$current_path/shell/zshrc ~/.zshrc"
+elif $REPLACE_FILES; then
   echo "    Deleting old zshrc!"
   rm ~/.zshrc
-  ln -sf $current_path/shell/zshrc ~/.zshrc
+  ln -sf "$current_path/shell/zshrc ~/.zshrc"
 else
   echo "    Keeping existing zshrc!"
 fi
@@ -85,11 +98,11 @@ echo -n "[ gitconfig ]"
 
 if [ ! -f ~/.gitconfig ]; then
   echo "    Creating gitconfig!"
-  ln -sf $current_path/git/gitconfig ~/.gitconfig
-else if REPLACE_FILES; then
+  ln -sf "$current_path/git/gitconfig ~/.gitconfig"
+elif REPLACE_FILES; then
   echo "    Deleting old gitconfig!"
   rm ~/.gitconfig
-  ln -sf $current_path/git/gitconfig ~/.gitconfig
+  ln -sf "$current_path/git/gitconfig ~/.gitconfig"
 else
   echo "    Keeping existing gitconfig!"
 fi
@@ -98,11 +111,11 @@ echo -n "[ gitignore ]"
 
 if [ ! -f ~/.gitignore ]; then
   echo "    Creating gitignore!"
-  ln -sf $current_path/git/gitignore ~/.gitignore
-else if REPLACE_FILES; then
+  ln -sf "$current_path/git/gitignore ~/.gitignore"
+elif REPLACE_FILES; then
   echo "    Deleting old gitignore!"
   rm ~/.gitignore
-  ln -sf $current_path/git/gitignore ~/.gitignore
+  ln -sf "$current_path/git/gitignore ~/.gitignore"
 else
   echo "    Keeping existing gitignore!"
 fi
@@ -112,7 +125,7 @@ fi
 #-----------------------------------------------------
 echo -n "[ Neovim ]"
 
-if !command_exists nvim; then
+if ! command_exists nvim; then
   echo "    Installing Neovim!"
   sudo add-apt-repository ppa:neovim-ppa/unstable
   sudo apt-get update
@@ -125,7 +138,7 @@ if [ ! -d ~/.config/nvim ]; then
   echo "    Creating nvim folder!"
   mkdir ~/.config/nvim
   install_nvim_folder
-else if REPLACE_FILES; then
+elif REPLACE_FILES; then
   echo "    Deleting old nvim folder!"
   rm -rf ~/.config/nvim
   install_nvim_folder
@@ -138,17 +151,17 @@ fi
 #-----------------------------------------------------
 echo -n "[ tmux.conf ]"
 
-if !command_exists tmux; then
+if ! command_exists tmux; then
   sudo apt-get install tmux -y
 fi
 
 if [ ! -f ~/.tmux.conf ]; then
   echo "    Creating tmux.conf!"
-  ln -sf $current_path/tmux/tmux.conf ~/.tmux.conf
-else if REPLACE_FILES; then
+  ln -sf "$current_path/tmux/tmux.conf ~/.tmux.conf"
+elif REPLACE_FILES; then
   echo "    Deleting old tmux.conf!"
   rm ~/.tmux.conf
-  ln -sf $current_path/tmux/tmux.conf ~/.tmux.conf
+  ln -sf "$current_path/tmux/tmux.conf ~/.tmux.conf"
 else
   echo "    Keeping existing tmux.conf!"
 fi
@@ -158,18 +171,18 @@ fi
 #-----------------------------------------------------
 echo -n "[ Xresources ]"
 
-if !command_exists xterm; then
+if ! command_exists xterm; then
   sudo apt-get install xterm -y
 fi
 
 if [ ! -f ~/.Xresources ]; then
   echo "   Creating Xresources!"
-  ln -sf $current_path/shell/Xresources ~/.Xresources
+  ln -sf "$current_path/shell/Xresources ~/.Xresources"
   xrdb -merge  ~/.Xresources
-else if REPLACE_FILES; then
+elif REPLACE_FILES; then
   echo "   Deleting old Xresources!"
   rm ~/.Xresources
-  ln -sf $current_path/shell/Xresources ~/.Xresources
+  ln -sf "$current_path/shell/Xresources ~/.Xresources"
   xrdb -merge  ~/.Xresources
 else
   echo "   Keeping existing Xresources!"
@@ -183,9 +196,9 @@ echo -n "[ Ruby (rbenv) and utilities (gemrc, irbrc, rdebugrc) ]"
 if command_exists ruby; then
   if [ ! -f ~/.gemrc ]; then
     echo "   Creating gemrc, irbrc, rdebugrc!"
-    ln -sf $current_path/ruby/gemrc ~/.gemrc
-    ln -sf $current_path/ruby/irbrc ~/.irbrc
-    ln -sf $current_path/ruby/rdebugrc ~/.rdebugrc
+    ln -sf "$current_path/ruby/gemrc ~/.gemrc"
+    ln -sf "$current_path/ruby/irbrc ~/.irbrc"
+    ln -sf "$current_path/ruby/rdebugrc ~/.rdebugrc"
   else
     echo "   Keeping existing gemrc, irbrc, rdebugrc!"
   fi
@@ -193,7 +206,7 @@ if command_exists ruby; then
   if command_exists pry; then
     if [ ! -f ~/.pryrc ]; then
       echo "   Creating pryrc!"
-      ln -sf $current_path/ruby/pryrc ~/.pryrc
+      ln -sf "$current_path/ruby/pryrc ~/.pryrc"
     else
       echo "   Keeping existing pryrc!"
     fi
@@ -201,7 +214,7 @@ if command_exists ruby; then
     echo "   Installing pry!"
     gem install pry
     echo "   Creating pryrc!"
-    ln -sf $current_path/ruby/pryrc ~/.pryrc
+    ln -sf "$current_path/ruby/pryrc ~/.pryrc"
   fi
 else
   echo "    Installing, rbenv and rubybuild."
@@ -220,7 +233,7 @@ echo -n "[ Ag ]"
 if command_exists ag; then
   if [ ! -f ~/.agignore ]; then
     echo "   Creating agignore!"
-    ln -sf $current_path/other/agignore ~/.agignore
+    ln -sf "$current_path/other/agignore ~/.agignore"
   else
     echo "   Keeping existing agignore!"
   fi
@@ -228,7 +241,7 @@ else
   echo "   Installing Ag!"
   sudo apt-get install -y silversearcher-ag
   echo "   Creating agignore!"
-  ln -sf $current_path/other/agignore ~/.agignore
+  ln -sf "$current_path/other/agignore ~/.agignore"
 fi
 
 #-----------------------------------------------------
@@ -241,21 +254,21 @@ if [ ! -d ~/dev ]; then
 fi
 
 if command_exists rubocop; then
-  ln -sf $current_path/linters/rubocop.yml ~/dev/.rubocop.yml
+  ln -sf "$current_path/linters/rubocop.yml ~/dev/.rubocop.yml"
 else
   echo "   Installing rubocop!"
   gem install rubocop
-  ln -sf $current_path/linters/rubocop.yml ~/dev/.rubocop.yml
+  ln -sf "$current_path/linters/rubocop.yml ~/dev/.rubocop.yml"
 fi
 
 echo -n "[ Eslint ]"
 
 if command_exists eslint; then
-  ln -sf $current_path/linters/eslintrc ~/dev/.eslintrc
+  ln -sf "$current_path/linters/eslintrc ~/dev/.eslintrc"
 else
   if command_exists npm; then
     npm install -g eslint
-    ln -sf $current_path/linters/eslintrc ~/dev/.eslintrc
+    ln -sf "$current_path/linters/eslintrc ~/dev/.eslintrc"
   else
     echo "   Install node and npm, then rerun script again!"
     exit
